requirements.txt
PyYaml==5.3.1
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
datefinder==0.7.1
html5lib==1.1
idna==2.10
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.13
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
requests==2.24.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.10
virtualenv==20.0.30
webencodings==0.5.1
zipp==3.1.0